requirements.txt
distro
requests
python-dotenv
pytest
psutil
docker
jinja2
rich
typer
pydantic-settings
